bugfix: restore v2 - 40k sampling rate and refactor conditional handling
Noticed that the code was making an exception with the line:
`if version19 == "v1" or sr2 == "40k":`
This approach isn't ideal, as it causes v2 - 40k to fallback to v1 - 40k. While this fix re-adds the necessary code, a better long-term solution might be to copy the v1 - 40k config into the v2 config folder to avoid such exceptions in the future.